refactor: remove pastas individuais dos componentes
