+++ src/components/MainLayout.jsx
@@ -1,4 +1,4 @@
-import Navbar from "../components/Navbar/Navbar";
+import Navbar from "./Navbar";
 
 const Layout = ({ children }) => {
   return (
