+++ src/routes/AppRoutes.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Route, BrowserRouter, Routes } from "react-router-dom";
-import GetData from '../components/MyRequests/MyRequests.jsx';
-import Navbar from "../components/Navbar/Navbar";
+import GetData from '../components/MyRequests.jsx';
+import Navbar from "../components/Navbar.jsx";
 import SignUp from "../pages/Citizen/SignUp.jsx";
 import HomeCitizen from "../pages/Citizen/HomeCitizen.jsx";
 import Login from "../pages/Citizen/Login.jsx";
